* Fixes to locally changed channel

+++ ESPDomotic.h
@@ -26,7 +26,7 @@ const uint8_t       _paramPortValueLength         = 6;    // port range is from 
 class Channel {
     public:
         Channel(const char* id, const char* name, uint8_t pin, uint8_t pinMode, uint8_t state);
-        Channel(const char* id, const char* name, uint8_t pin, uint8_t pinMode, uint8_t state, uint16_t timer);
+        Channel(const char* id, const char* name, uint8_t pin, uint8_t pinMode, uint8_t state, uint32_t timer);
 
         const char*     id;
         char*           name;
@@ -39,7 +39,7 @@ class Channel {
 
         unsigned long   timerControl;
         
-        void    init(const char* id, const char* name, uint8_t pin, uint8_t pinMode, uint8_t state, uint16_t timer);
+        void    init(const char* id, const char* name, uint8_t pin, uint8_t pinMode, uint8_t state, uint32_t timer);
 
         // Updates the channel´s name
         void    updateName (const char *v);
